refactor: Migrate consent form to mvvm
