+++ lib/src/routes.dart
@@ -2,7 +2,7 @@ import 'package:get/get.dart';
 import 'package:mdigits/src/demographics/close_demographics_screen.dart';
 import 'package:mdigits/src/demographics/end_demographics_survey.dart';
 import 'package:mdigits/src/ema/ema_screen.dart';
-import 'package:mdigits/src/informed_consent/consent_screen.dart';
+import 'package:mdigits/src/informed_consent/view/consent_page.dart';
 import 'package:mdigits/src/notifications/view/notification_permission_screen.dart';
 import 'package:mdigits/src/task_list/view/pages/task_list_page.dart';
 import 'package:mdigits/src/ui_components/loading_screen.dart';
@@ -11,7 +11,7 @@ import 'surveys/demographics_survey.dart';
 
 final List<GetPage> routes = <GetPage>[
   GetPage(name: '/tasklist', page: () => TaskListPage()),
-  GetPage(name: '/consent', page: () => ConsentScreen()),
+  GetPage(name: '/consent', page: () => ConsentPage()),
   GetPage(
       name: '/demographicsSurvey',
       page: () => DemographicsSurvey(
